Preliminary support for Satisfactory 1.0 plus additions
A new variant of TypedData called "ClientIdentifyInfo" needs to be processed. It contains a hex string as an ID, along with some additional but currently unknown bytes.
The property "mDestroyedPickups" also requires processing. It is currently treated as a collection of Vector4s, though this may or may not be accurate and requires further clarification.
The list of items with state (i.e., "FueledInventoryItem") has been updated to include the object scanner and the gas mask. It may be worth considering renaming the subclass, e.g., to "StatefulInventoryItem".